Update README and all the scripts to reproduce the results

+++ include/FedTree/common.h
@@ -42,7 +42,7 @@ std::string string_format(const std::string &format, Args ... args) {
 }
 
 //data types
-typedef float float_type;
+typedef double float_type;
 
 #define EPSILON 1e-9
